virtual_sound_card/requirements.txt
daily
numpy
pyaudio
pyvirtualcam
python-dotenv
requests
scipy
sounddevice
vapi_python 